# feat: Switch to Castagnoli Polynomial for Faster CRC32 Computation