replaces the default IEEE polynomial with the Castagnoli polynomial (crc32.Castagnoli) for CRC32 checksum calculations.

Running this on fuzzer with 1000 local relayer and pprof.

## Before
```
4.72s  3.83% 92.87%      4.72s  3.83%  hash/crc32.ieeeUpdate
```

## After
```
 2.67s  2.14% 90.29%      2.67s  2.14%  hash/crc32.ieeeUpdate
```

This comes roughly as 43.33% faster in our WAL.